Saving data
some more validation and then we use
the Mongo method insert to save our data. The first parameter is a hash
of key:values to save. We then have an optional second parameter, in
which Mongo gives you two parameters to deal with (if you want to!), err
and id. If the insert() was successful, then id is the id of the
inserted record, if it wasn’t successful then err contains the error
from Mongo. Depending on the results of your if statements, it’s upto
you how you deal with that in the UI. I went for a classy alert() if it
all went belly up